Move away from module in node.js
