+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "description": "The project repository for an open system to generate and manage certificates.",
   "main": "index.js",
-  "type": "module",
   "scripts": {
     "prettier": "npx prettier --write .",
     "prettier:check": "npx prettier --check .",
