+++ src/config/cloudinary.js
@@ -1,6 +1,6 @@
-import { v2 as cloudinary } from 'cloudinary';
+const { v2: cloudinary } = require('cloudinary');
 
-import logger from './winston.js';
+const logger = require('./winston.js');
 
 try {
   cloudinary.config({
