+++ src/graphql/mutation/index.js
@@ -2,17 +2,14 @@
 import { GraphQLObjectType } from 'graphql';
 
 // Mutations
-import { addNewWelcomeMessage, deleteWelcomeMessage } from './welcome.js';
-import { createUser, updateUser } from './user.js';
+const { addNewWelcomeMessage, deleteWelcomeMessage } = require('./welcome.js');
 
 const Mutation = new GraphQLObjectType({
   name: 'mutation',
   fields: () => ({
     addNewWelcomeMessage,
     deleteWelcomeMessage,
-    createUser,
-    updateUser,
   }),
 });
 
-export default Mutation;
+module.exports = Mutation;
