--- src/graphql/types/index.js
@@ -1,4 +1,0 @@
-export { default as WelcomeType } from './Welcome.js';
-export { default as RequestType } from './RequestType.js';
-export { default as UserType } from './UserType.js';
-export { default as SignType } from './SignType.js';
