--- src/models/index.js
@@ -1,5 +1,0 @@
-export { default as Certificate } from './certificate.js';
-export { default as Request } from './request.js';
-export { default as User } from './user.js';
-export { default as Sign } from './sign.js';
-export { default as Welcome } from './welcome.js';
